imap: move over imapclient.BodyStructure
